Refactor: Separate `PDAG`'s test code from `DAG`. (#2385)
